tightening up markup
took out attributes from footer and header the nav. and instead added
divs inside with the matching attributes there parents had. Changed all
the attributes that had underscores to dashes // and matched the labels
for attributes to there matching id’s.
    Gave the markup better S curves.
 changed the wrap id and added a class attribute for the clear fix .. i
also did the same for the elements the had more than one class but was
missing a
separation .